% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{addYear}
\alias{addYear}
\title{Add Next Year to the year-by-year RIME data}
\usage{
addYear(df)
}
\arguments{
\item{df}{Unstructured year-by-year RIME data of one response from the
survey}
}
\value{
Structured year-by-year data with next year added to the final
column. Also writes this to a .csv file so the data can be added later.
}
\description{
Add Next Year to the year-by-year RIME data
}
