% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{chiSqTest}
\alias{chiSqTest}
\title{Pearson's Chi Square Test}
\usage{
chiSqTest(df, var1, var2, idvar = ID)
}
\arguments{
\item{df}{Data frame of one year RIME data}

\item{var1}{Variable 1}

\item{var2}{Variable 2}

\item{idvar}{Identifying variable}
}
\value{
p-value of the Pearson Chi-Square test
}
\description{
Pearson's Chi Square Test
}
\examples{
totDf <- read.csv("Data Info/2021 IMEB Data.csv")
agegrp2 <- cut(totDf$Age, c(24,29,44), labels = c("<30","30+"))
totDf$AgeGroup2 <- agegrp2

chiSqTest(df = totDf, var1 = AgeGroup2, var2 = Q25a_balance, idvar = ID)
}
