% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{contTbl}
\alias{contTbl}
\title{2 Dimensional Contingency Table}
\usage{
contTbl(df, var1, var2)
}
\arguments{
\item{df}{Data frame of one year RIME data}

\item{var1}{Variable along the top that will be shown in the output}

\item{var2}{Variable along the side that will not be shown}
}
\value{
2 Dimensional Contingency Table
}
\description{
2 Dimensional Contingency Table
}
\examples{
totDf <- read.csv("Data Info/2021 IMEB Data.csv")
agegroups5 <- c("25-29", "30-34", "35–39", "40-44", "45+")
agegrp5 <-cut(totDf$Age, c(24,29,34,39,44,100),
              labels = agegroups5)
totDf$AgeGroup <- agegrp5

contTbl(df = totDf, var1 = AgeGroup, var2 = Q25a_balance)
}
