% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{descStats}
\alias{descStats}
\title{Totals or average by response}
\usage{
descStats(df, stat = c("total", "average"))
}
\arguments{
\item{df}{unstructured data frame}

\item{stat}{choice of either c("total", "average")}
}
\value{
structured data frame with descriptive statistic column
}
\description{
Totals or average by response
}
