% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{normalize}
\alias{normalize}
\title{Normalize Year-by-year data}
\usage{
normalize(df)
}
\arguments{
\item{df}{Unstructured year-by-year RIME data of one response from the
survey}
}
\value{
List of responses possible responses and structured data with
only the
}
\description{
Normalize Year-by-year data
}
