% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFunctions.R
\name{separate}
\alias{separate}
\title{Separate Function}
\usage{
separate(df, sep_by)
}
\arguments{
\item{df}{Data frame of one year RIME data}

\item{sep_by}{Variable to separate by}
}
\value{
List of Data Frames by the possible Responses of sep_by
}
\description{
Separate Function
}
\examples{
totDf <- read.csv("Data Info/2021 IMEB Data.csv")

dfList <- separate(totDf, Hospitalcode)

}
